Use lerna to build multiple modules
