+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
     "files.exclude": {
         "**/node_modules": true,
+        "**/.nyc_output": true
     },
     "xo.enable": true
 }
