+++ lerna.json
@@ -1,0 +1,7 @@
+{
+  "packages": [
+    "packages/*"
+  ],
+  "npmClientArgs": ["--no-package-lock"],
+  "version": "0.0.8"
+}
