+++ packages/livedata/index.test-d.ts
@@ -1,5 +1,5 @@
 import {expectType} from 'tsd'
-import {LiveData, MediatorLiveData, map, switchMap} from '.'
+import {LiveData, MediatorLiveData} from '.'
 
 expectType<unknown>(new LiveData().get())
 expectType<boolean>(new LiveData(true).get())
@@ -10,6 +10,3 @@ expectType<void>(new LiveData(true).subscribe(() => true)())
 expectType<boolean>(new MediatorLiveData(true).get())
 expectType<unknown>(new MediatorLiveData().get())
 expectType<() => void>(new MediatorLiveData().addSource(new LiveData(), v => {}))
-
-expectType<LiveData<string>>(map(new LiveData(true), v => 'string'))
-expectType<LiveData<string>>(switchMap(new LiveData(true), v => new LiveData<string>('string')))
