+++ packages/operators/index.js
@@ -1,0 +1,1 @@
+export {switchMap, map} from './transformations'
